feat: Make XCOM push optional with result dataclass support
The dbt operators will now avoid returning anything if xcom_push is
set to False (default). This mimics the behavior found in Airflow's
BashOperator.

Moreover, some dbt operations return special result dataclasses which
are not serializable. This was causing issues with Airflow's XCOM
feature. When we see a dataclass, we now transform it to a dict before
returning (if xcom_push is set to True).