fix: Serialize agate tables in dbt seed results
dbt seed command returns an agate.table.Table as part of its
results. This object is not json serializable by default so
DbtSeedOperator was failing whenever it ran with do_xcom_push. We now
handle agate tables by turning them into dictionaries of column names
and types, which should be enough information. We could have included
the full table as a json dict, but that seemed a bit excessive.

Unit tests were extended to ensure json.dumps does succeed when run on
the results of operators with do_xcom_push set to True