test: Add testing support for dbt version 0.20
dbt version 0.20 should be supported by the code as it is, going by
unit tests passing with minor changes:
* A couple of dbt commands now return booleans instead of None
* A dispatch entry in dbt_project.yml is required

Regardless, support for 0.20 should be considered experimental until
we've had more time to play around with it. Make sure to pin your dbt
version to any v0.19 patch version if you prefer to opt out of v0.20.