feat(operator): Add new operator for dbt docs generate
Support for dbt docs generate itself is pretty trivial as we only needed a
new configuration and dbt docs generate does not take many
configuration arguments. The biggest challenge of supporting dbt docs generate
is solving the issue of running dbt docs generate in a temporary
directory: there isn't much sense in executing dbt docs generate if we
cannot use the results of the execution. The alternative of pushing
the dbt artifacts to XCom is, of course, still available, but I wanted
to use this an excuse to develop something else.

So, we also added a push_dbt_project method to our S3
hook. This method is called right before exiting the temporary
directory, and can be controlled with a push_dbt_project argument to
the operator. The argument is by default True for the new
DbtDocsGenerateOperator so that we can push the documentation files to
S3. It is also True by default when running DbtDepsOperator since
there was no way for us install dependencies.

In the future, I'd like us to support more "dbt backends" besides S3:
maybe support for pulling projects from Github? SSH and SFTP
connections? Lots of possibilities!