feat(test): Test project files when using DptDepsOperator
DbtDepsOperator could have unintended side-effects if it were to
modify files outside of the packages directory. For this reason, we
added a new test, and extended existing tests to ensure no files are
changed in a dbt project outside of the dbt packages directory.

This should provide more guarantees about the behavior of the
operator, specially when looking to push back files to S3.