fix: Set results before running task to avoid UnboundLocalError
dbt handles exceptions raised by task.run() with the track_run context
manager. We are also doing the same thing. The problem is that some
exceptions, like FailedToConnectException, are failing silently (at
least for us, CLI users do see a log) and are not re-raised by
dbt.

Continuing with execution would lead to an UnboundLocalError as
we don't account for results never being set. To address this we
preemptively set results to None before running a task.