feat: Support Airflow connections for dbt targets
The objective of airflow-dbt-python has always been to make dbt a
first-class citizen of Airflow. As part of this goal, we want to
integrate dbt with all of Airflow's features. In particular, Airflow
connections allow us to safely store connection information for
operators to use, so all dbt operators should be able to leverage them
too.

The way we achieve this is by manually instantiating a dbt Project and
Profile. When reading the latter, we also inject any Airflow
connections that match the given target argument (if any). Moreover,
if the profile is not defined, we simply create our own with any
Airflow connection that was passed as a target (of course, missing
both a profiles file and an Airflow connection raises an error).

Future work may extend this to support custom connection types, at the
moment we are doing a best effort to include all possible arguments,
but it's not perfect.

This feature was suggested in the dbt Slack channel as a way to avoid
having to manage a profiles.yml file, that may contain sensitive
information, making it a bad target for version control.